refactor: simplify C codegen and unify scalar types to i64
- Replace complex scheduling annotations with flat per-element OpenMP loop generation
- Change scalar integer types from i32 to i64 throughout the codebase
- Remove macOS-specific compiler paths and simplify compilation process

+++ crates/sile/examples/softmax.rs
@@ -1,16 +1,16 @@
-use sile::{Device, Tensor, Tile};
+use sile::{Device, Tensor};
 
 #[sile::kernel]
-fn softmax<const BM: i32, const BN: i32>(
+fn softmax<const BM: i64, const BN: i64>(
     x: &Tensor<f32, { [-1, -1] }>,
     y: &mut Tensor<f32, { [BM, BN] }>,
 ) {
     let tile_x: Tile<f32, { [BM, BN] }> = sile::load_tile_like_2d(x, y);
-    let tile_x_max: Tile<f32, { [BM] }> = sile::reduce_max(tile_x.clone(), 1i32);
+    let tile_x_max: Tile<f32, { [BM] }> = sile::reduce_max(tile_x.clone(), 1i64);
     let tile_x_max_bcast: Tile<f32, { [BM, BN] }> =
         tile_x_max.reshape([BM, 1]).broadcast(y.shape());
     let num: Tile<f32, { [BM, BN] }> = sile::exp(tile_x - tile_x_max_bcast);
-    let denom: Tile<f32, { [BM] }> = sile::reduce_sum(num.clone(), 1);
+    let denom: Tile<f32, { [BM] }> = sile::reduce_sum(num.clone(), 1i64);
     let denom_bcast: Tile<f32, { [BM, BN] }> = denom.reshape([BM, 1]).broadcast(y.shape());
     y.store(num / denom_bcast);
 }
